+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Nothing yet
+
+## [0.3.0] - 2025-10-14
+
 ### Added
 
 - New `mp2ts-extract` tool to extract elementary video streams from TS files
@@ -47,7 +51,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - initial version of the repo
 - ts-info tool
 
-[Unreleased]: https://github.com/Eyevinn/mp2ts-tools/releases/tag/v0.2.1...HEAD
+[Unreleased]: https://github.com/Eyevinn/mp2ts-tools/releases/tag/v0.3.0...HEAD
+[0.3.0]: https://github.com/Eyevinn/mp2ts-tools/releases/tag/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/Eyevinn/mp2ts-tools/releases/tag/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/Eyevinn/mp2ts-tools/releases/tag/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/Eyevinn/mp2ts-tools/releases/tag/v0.1.0